refactor: features 폴더 리펙토링 (#84)
