+++ src/features/health/ui/GraphSection.tsx
